pyproject.toml
[tool.poetry]
name = "int-shop"
version = "0.1.0"
description = ""
authors = ["mitsy01 <alexanderfigmochka19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-restful = "^0.3.10"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.37"
pg8000 = "^1.31.2"
requests-html = "^0.10.0"
lxml = "^5.3.0"
flask-sqlalchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
